chore(cli): refactor script to extract and save versions of all SB packages
The previous script was only extracting version for packages under `app`
directory. Now, all `@storybook` packages are processed and save with
their current version in the monorepo in `versions.json`. This file will
be use to find out which version of the packages should be installed
when running `sb init`.