feat(cli): use `versions.json` to get SB packages versions
Previously it used `devDep` of `package.json`. It was working but doing
so all Storybook packages had to be added to `devDep`.

Moreover, as `@storybook/react-native` was not part of the monorepo
anymore it's version was stuck to `6.0.0-alpha.0` causing a download
of `6.0.0-alpha.0` SB packages when running `yarn install`.